+++ src/Agents.jl
@@ -12,8 +12,10 @@ import LinearAlgebra
 
 # Core structures of Agents.jl
 include("core/agents.jl")
-include("core/model.jl")
+include("core/model_abstract.jl")
+include("core/model_concrete.jl")
 include("core/space_interaction_API.jl")
+include("core/higher_order_iteration.jl")
 
 # Existing spaces
 include("spaces/nothing.jl")
@@ -47,29 +49,26 @@ using Scratch
 
 function __init__()
 display_update = true
-version_number = "5.5"
+version_number = "5.7"
 update_name = "update_v$(version_number)"
 update_message = """
 Update message: Agents v$(version_number)
 Welcome to this new update of Agents.jl!
+
 Noteworthy changes:
 
-- The `@agent` macro has been re-written and is now more general and more safe.
-  It now also allows inhereting fields from any other type.
-- The `@agent` macro is now THE way to create agent types for Agents.jl simulations.
-  Directly creating structs by hand is no longer mentioned in the documentation at all.
-  This will allow us in the future to utilize additional fields that the user does not
-  have to know about, which may bring new features or performance gains by being
-  part of the agent structures.
-- The minimal agent types like `GraphAgent` can be used normally as standard agent
-  types that only have the mandatory fields. This is now clear in the docs.
-  (this was possible also before v5.4, just not clear)
-- In the future, making agent types by hand may be completely dissalowed, resulting
-  in error. Therefore, making agent types manually is considered deprecated.
-- New function `normalize_position`.
-
-See the CHANGELOG.md, because v5.4 was also a large release!
+- Internals of `AgentBasedModel` got reworked.
+  It is now an abstract type, defining an abstract interface that concrete
+  implementations may satisfy. This paves the way for flexibly defining new
+  variants of `AgentBasedModel` that are more specialized in their applications.
+- The old `AgentBasedModel` is now `StandardABM`
+  Nothing is breaking because the call to `AgentBasedModel` gives `StandardABM`.
+- Two new variants of agent based models: `UnkillableABM` and `FixedMassABM`.
+  They yield huge performance benefits **(up to twice the speed!!!)**
+  on iterating over agents if the agents can't get killed, or even added,
+  during model evolution!
 """
+
 if display_update
     # Get scratch space for this package
     versions_dir = @get_scratch!("versions")