+++ src/core/agents.jl
@@ -2,18 +2,18 @@ export AbstractAgent, @agent, NoSpaceAgent
 
 """
     YourAgentType <: AbstractAgent
+
 Agents participating in Agents.jl simulations are instances of user-defined Types that
 are subtypes of `AbstractAgent`.
 
 Your agent type(s) **must have** the `id::Int` field as first field.
-In Julia versions ≥ v1.8, this must also be declared as a `const` field.
 If any space is used (see [Available spaces](@ref)), a `pos` field of appropriate type
 is also mandatory. The core model structure, and each space,
 may also require additional fields that may,
 or may not, be communicated as part of the public API.
 
 The [`@agent`](@ref) macro ensures that all of these constrains are in place
-and hence it is the only officially supported way to generate new agent types.
+and hence it is the recommended way to generate new agent types.
 """
 abstract type AbstractAgent end
 
@@ -28,12 +28,17 @@ Define an agent struct which includes all fields that `AnotherAgentType` has,
 as well as any additional ones the user may provide via the `begin` block.
 See below for examples.
 
-Using `@agent` is **the only supported way to create agent types** for Agents.jl.
+Using `@agent` is the recommended way to create agent types for Agents.jl,
+however keep in mind that the macro (currently) doesn't work with `Base.@kwdef`
+or `const` declerations in individual fields (for Julia v1.8+).
+
 Structs created with `@agent` by default subtype `AbstractAgent`.
 They cannot subtype each other, as all structs created from `@agent` are concrete types
 and `AnotherAgentType` itself is also concrete (only concrete types have fields).
 If you want `YourAgentType` to subtype something other than `AbstractAgent`, use
 the optional argument `OptionalSupertype` (which itself must then subtype `AbstractAgent`).
+
+## Usage
 
 The macro `@agent` has two primary uses:
 1. To include the mandatory fields for a particular space in your agent struct.
@@ -49,11 +54,8 @@ The existing minimal agent types are:
 
 All will attribute an `id::Int` field, and besides `NoSpaceAgent` will also attribute
 a `pos` field. You should **never directly manipulate the mandatory fields `id, pos`**
-that the resulting new agent type will have.
-The `id` is an unchangable field (and in Julia versions ≥ v1.8 this is enforced).
+that the resulting new agent type will have. The `id` is an unchangable field.
 Use functions like [`move_agent!`](@ref) etc., to change the position.
-
-You can use the `@doc` macro from Julia to document the generated struct if you wish so.
 
 ## Examples
 ### Example without optional hierarchy
@@ -89,9 +91,9 @@ mutable struct Baker{T} <: AbstractAgent
     breadz_per_day::T
 end
 ```
-### Example with optional hierachy
+### Example with optional hierarchy
 An alternative way to make the above structs, that also establishes
-a user-specific subtyping hierachy would be to do:
+a user-specific subtyping hierarchy would be to do:
 ```julia
 abstract type AbstractHuman <: AbstractAgent end
 
@@ -167,41 +169,6 @@ f(x::Animal) = ... # uses `CommonTraits` fields
 f(x::Person) = ... # uses fields that all "persons" have
 ```
 """
-macro agent(new_name, base_type, extra_fields)
-    # This macro was generated with the guidance of @rdeits on Discourse:
-    # https://discourse.julialang.org/t/
-    # metaprogramming-obtain-actual-type-from-symbol-for-field-inheritance/84912
-
-    # We start with a quote. All macros return a quote to be evaluated
-    quote
-        let
-            # Here we collect the field names and types from the base type
-            # Because the base type already exists, we escape the symbols to obtain it
-            base_fieldnames = fieldnames($(esc(base_type)))
-            base_fieldtypes = [t for t in getproperty($(esc(base_type)), :types)]
-            base_fields = [:($f::$T) for (f, T) in zip(base_fieldnames, base_fieldtypes)]
-            # Then, we prime the additional name and fields into QuoteNodes
-            # We have to do this to be able to interpolate them into an inner quote.
-            name = $(QuoteNode(new_name))
-            additional_fields = $(QuoteNode(extra_fields.args))
-            # Now we start an inner quote. This is because our macro needs to call `eval`
-            # However, this should never happen inside the main body of a macro
-            # There are several reasons for that, see the cited discussion at the top
-            expr = quote
-                mutable struct $name <: AbstractAgent
-                    $(base_fields...)
-                    $(additional_fields...)
-                end
-            end
-            # @show expr # uncomment this to see that the final expression looks as desired
-            # It is important to evaluate the macro in the module that it was called at
-            Core.eval($(__module__), expr)
-        end
-    end
-end
-# TODO: I do not know how to merge these two macros to remove code duplication.
-# There should be away that only the 4-argument version is used
-# and the 3-argument version just passes `AbstractAgent` to the 4-argument.
 macro agent(new_name, base_type, super_type, extra_fields)
     # This macro was generated with the guidance of @rdeits on Discourse:
     # https://discourse.julialang.org/t/
@@ -234,15 +201,28 @@ macro agent(new_name, base_type, super_type, extra_fields)
             # It is important to evaluate the macro in the module that it was called at
             Core.eval($(__module__), expr)
         end
+        # allow attaching docstrings to the new struct, issue #715
+        Core.@__doc__($(esc(Docs.namify(new_name))))
+        nothing
     end
 end
 
+macro agent(new_name, base_type, extra_fields)
+    # Here we nest one macro call into another because there is no way to provide 
+    # defaults for macro arguments. We proceed to call the actual macro with the default
+    # `super_type = AbstractAgent`. This requires us to disable 'macro hygiene', see here
+    # for a brief explanation of the potential issues with this: 
+    # https://discourse.julialang.org/t/calling-a-macro-from-within-a-macro-revisited/19680/16?u=fbanning
+    esc(quote
+        Agents.@agent($new_name, $base_type, Agents.AbstractAgent, $extra_fields)
+    end)
+end
 
 """
     NoSpaceAgent <: AbstractAgent
 The minimal agent struct for usage with `nothing` as space (i.e., no space).
 It has the field `id::Int`, and potentially other internal fields that
-are not documentated as part of the public API. See also [`@agent`](@ref).
+are not documented as part of the public API. See also [`@agent`](@ref).
 """
 mutable struct NoSpaceAgent <: AbstractAgent
     id::Int