+++ test/model_access.jl
@@ -1,15 +1,16 @@
+using Agents, Test
+
 @testset "Model Access" begin
 @testset "Accessing model" begin
-    model = ABM(Agent0; properties = Dict(:a => 2, :b => "test"))
+    @testset "ModelType=$(ModelType)" for ModelType in (ABM, StandardABM, UnkillableABM)
+    model = ModelType(NoSpaceAgent; properties = Dict(:a => 2, :b => "test"))
     for i in 1:5
         add_agent!(model)
     end
-    @test model.scheduler == Schedulers.fastest
-    @test typeof(model.agents) <: Dict
-    @test model.space === nothing
-    @test model.properties == Dict(:a => 2, :b => "test")
+    @test abmscheduler(model) == Schedulers.fastest
+    @test abmproperties(model) == Dict(:a => 2, :b => "test")
     a = model[1]
-    @test a isa Agent0
+    @test a isa NoSpaceAgent
     @test a.id == 1
     @test model.a == 2
     @test model.b == "test"
@@ -20,14 +21,11 @@
     @test model.b == "Changed"
     @test_throws ErrorException model.c = 5
 
-    newa = Agent0(6)
-    model[6] = newa
+    newa = NoSpaceAgent(6)
+    add_agent!(newa, model)
     @test model[6] == newa
-    @test_throws ArgumentError model[7] = newa
-
-    prop2 = Agent2(1, 0.5)
-    model2 = ABM(Agent0; properties = prop2)
-    @test model2.weight == 0.5
+    # setindex must errors
+    @test_throws ErrorException model[7] = newa
 
     mutable struct Properties
         par1::Int
@@ -35,7 +33,8 @@
         par3::String
     end
     properties = Properties(1,1.0,"Test")
-    model = ABM(Agent0; properties = properties)
+    model = ModelType(NoSpaceAgent; properties)
+    @test abmproperties(model) isa Properties
     @test model.par1 == 1
     @test model.par2 == 1.0
     @test model.par3 == "Test"
@@ -48,15 +47,16 @@
     @test model.par3 == "Changed"
     @test_throws ErrorException model.par4 = 5
 
-    model = ABM(Agent0)
+    model = ModelType(NoSpaceAgent)
     @test_throws ErrorException model.a = 5
+    end
 end
 
 @testset "model access typestability" begin
     prop1 = Dict(:a => 0.5)
     prop2 = Agent2(1, 0.5)
-    model1 = ABM(Agent0; properties = prop1)
-    model2 = ABM(Agent0; properties = prop2)
+    model1 = ABM(NoSpaceAgent; properties = prop1)
+    model2 = ABM(NoSpaceAgent; properties = prop2)
 
     test1(model) = model.a
     test2(model) = model.weight
@@ -73,15 +73,15 @@ end
 end
 
 @testset "Display" begin
-    model = ABM(Agent0)
+    model = ABM(NoSpaceAgent)
     @test occursin("no spatial structure", sprint(show, model))
     model = ABM(Agent3, GridSpace((5,5)))
-    @test sprint(show, model)[1:29] == "AgentBasedModel with 0 agents"
-    @test sprint(show, model.space) == "GridSpace with size (5, 5), metric=chebyshev, periodic=true"
+    @test sprint(show, model)[1:25] == "StandardABM with 0 agents"
+    @test sprint(show, Agents.abmspace(model)) == "GridSpace with size (5, 5), metric=chebyshev, periodic=true"
     model = ABM(Agent6, ContinuousSpace((1.0,1.0)))
-    @test sprint(show, model.space) == "periodic continuous space with (1.0, 1.0) extent and spacing=0.05"
+    @test sprint(show, Agents.abmspace(model)) == "periodic continuous space with (1.0, 1.0) extent and spacing=0.05"
     model = ABM(Agent5, GraphSpace(path_graph(5)))
-    @test sprint(show, model.space) == "GraphSpace with 5 positions and 4 edges"
+    @test sprint(show, Agents.abmspace(model)) == "GraphSpace with 5 positions and 4 edges"
 end
 
 end